requirements.txt
ai2thor==5.0.0
openai==1.52.2
botocore==1.34.162
numpy==1.26.4
scipy==1.11.4
pillow==10.4.0
jsonlines==4.0.0
